consolidate some utils and types
