+++ legate/rc.py
@@ -14,14 +14,6 @@
 #
 from __future__ import annotations
 
-import sys
-import warnings
-from argparse import ArgumentParser, Namespace
-from dataclasses import dataclass, fields
-from typing import Any, Iterable, Literal, Sequence, Type, TypeVar, Union
-
-from typing_extensions import TypeAlias
-
 LEGION_WARNING = """
 
 All Legate programs must be run with a legion_python interperter. We
@@ -55,96 +47,3 @@ def check_legion(msg: str = LEGION_WARNING) -> None:
     """Raise an error if we are not running in legion_python."""
     if not has_legion_context():
         raise RuntimeError(msg)
-
-
-class _UnsetType:
-    pass
-
-
-Unset = _UnsetType()
-
-_T = TypeVar("_T")
-NotRequired = Union[_UnsetType, _T]
-
-
-def entries(obj: Any) -> Iterable[tuple[str, Any]]:
-    for f in fields(obj):
-        value = getattr(obj, f.name)
-        if value is not Unset:
-            yield (f.name, value)
-
-
-# https://docs.python.org/3/library/argparse.html#action
-ActionType: TypeAlias = Literal[
-    "store",
-    "store_const",
-    "store_true",
-    "append",
-    "append_const",
-    "count",
-    "help",
-    "version",
-    "extend",
-]
-
-# https://docs.python.org/3/library/argparse.html#nargs
-NargsType: TypeAlias = Literal["?", "*", "+", "..."]
-
-
-@dataclass(frozen=True)
-class ArgSpec:
-    dest: str
-    action: NotRequired[ActionType] = "store_true"
-    nargs: NotRequired[Union[int, NargsType]] = Unset
-    const: NotRequired[Any] = Unset
-    default: NotRequired[Any] = Unset
-    type: NotRequired[Type[Any]] = Unset
-    choices: NotRequired[Sequence[Any]] = Unset
-    help: NotRequired[str] = Unset
-    metavar: NotRequired[str] = Unset
-
-
-@dataclass(frozen=True)
-class Argument:
-    name: str
-    spec: ArgSpec
-
-
-def parse_command_args(libname: str, args: Iterable[Argument]) -> Namespace:
-    """ """
-    if not libname.isidentifier():
-        raise ValueError(
-            f"Invalid library {libname!r} for command line arguments"
-        )
-
-    parser = ArgumentParser(
-        prog=f"<{libname} program>", add_help=False, allow_abbrev=False
-    )
-
-    lib_prefix = f"-{libname}:"
-
-    argnames = [arg.name for arg in args]
-
-    for arg in args:
-        argname = f"{lib_prefix}{arg.name}"
-        kwargs = dict(entries(arg.spec))
-        parser.add_argument(argname, **kwargs)
-
-    has_custom_help = "help" in argnames
-
-    if f"{lib_prefix}help" in sys.argv and not has_custom_help:
-        parser.print_help()
-        sys.exit()
-
-    args, extra = parser.parse_known_args()
-
-    for item in extra:
-        if item.startswith(lib_prefix):
-            warnings.warn(
-                f"Unrecognized argument {item!r} for {libname} (passed on as-is)"  # noqa: E501
-            )
-            break
-
-    sys.argv = sys.argv[:1] + extra
-
-    return args
