+++ tests/unit/legate/utils/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2021-2022 NVIDIA Corporation
+# Copyright 2022 NVIDIA Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,13 +13,3 @@
 # limitations under the License.
 #
 from __future__ import annotations
-
-from typing import Any
-
-from typing_extensions import TypeAlias
-
-GenConfig: TypeAlias = Any
-
-GenSystem: TypeAlias = Any
-
-GenObjs: TypeAlias = Any
